switch to floats

+++ src/input_stdin.cpp
@@ -7,15 +7,15 @@
 void init_input() {
 }
 
-int16_t data[num_channels];
+float data[num_channels];
 
-const int16_t* input() {
+const float* input() {
 	while (true) {
 		while (getchar() != '\n');
 
 		bool failed = false;
 		for (size_t i = 0; i < num_channels; ++i)
-			failed = failed || scanf(" %" SCNd16, &data[i]) != 1;
+			failed = failed || scanf(" %f", &data[i]) != 1;
 
 		if (!failed) return data;
 	}
